requirements.txt
opencv-python
numpy
scikit-learn
joblib
tqdm
streamlit
Pillow
tensorflow 
matplotlib